edit UserSlice: convert to tsx
