--- shop-app/src/redux/UserSlice.jsx
@@ -1,19 +1,0 @@
-import { createSlice } from "@reduxjs/toolkit";
-
-export const UserSlice = createSlice({
-    name: 'user',
-    initialState: {
-        isLoggedIn: false,
-    },
-    reducers: {
-        logIn: (state) => {
-            state.isLoggedIn = true;
-        },
-        logOut: (state) => {
-            state.isLoggedIn = false;
-        }
-    }
-})
-
-export const { logIn, logOut } = UserSlice.actions;
-export default UserSlice.reducer;
